Move the storage interfaces to one folder
